# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  gymtrack-backend_internal_domain_models.FoodItem:
    properties:
      calories:
        minimum: 0
        type: number
      food:
        type: string
      macros:
        $ref: '#/definitions/gymtrack-backend_internal_domain_models.Macros'
      quantity:
        type: string
    required:
    - food
    - quantity
    type: object
  gymtrack-backend_internal_domain_models.Macros:
    properties:
      carbs:
        minimum: 0
        type: number
      fats:
        minimum: 0
        type: number
      protein:
        minimum: 0
        type: number
    type: object
  gymtrack-backend_internal_domain_models.Meal:
    properties:
      athleteId:
        type: string
      createdAt:
        type: string
      date:
        type: string
      items:
        items:
          $ref: '#/definitions/gymtrack-backend_internal_domain_models.FoodItem'
        minItems: 1
        type: array
      mealId:
        type: string
      mealType:
        allOf:
        - $ref: '#/definitions/gymtrack-backend_internal_domain_models.MealType'
        enum:
        - breakfast
        - lunch
        - dinner
        - snack
      type:
        description: Always "meal"
        type: string
      updatedAt:
        type: string
    required:
    - athleteId
    - date
    - items
    - mealType
    type: object
  gymtrack-backend_internal_domain_models.MealType:
    enum:
    - breakfast
    - lunch
    - dinner
    - snack
    type: string
    x-enum-varnames:
    - MealTypeBreakfast
    - MealTypeLunch
    - MealTypeDinner
    - MealTypeSnack
...
  internal_api_handlers.CreateMealRequest:
    description: Request body for creating a new meal entry
    properties:
      date:
        type: string
      items:
        items:
          $ref: '#/definitions/gymtrack-backend_internal_domain_models.FoodItem'
        minItems: 1
        type: array
      mealType:
        allOf:
        - $ref: '#/definitions/gymtrack-backend_internal_domain_models.MealType'
        enum:
        - breakfast
        - lunch
        - dinner
        - snack
    required:
    - date
    - items
    - mealType
    type: object
...
  internal_api_handlers.UpdateMealRequest:
    description: Request body for updating an existing meal entry
    properties:
      date:
        type: string
      items:
        items:
          $ref: '#/definitions/gymtrack-backend_internal_domain_models.FoodItem'
        minItems: 1
        type: array
      mealType:
        allOf:
        - $ref: '#/definitions/gymtrack-backend_internal_domain_models.MealType'
        enum:
        - breakfast
        - lunch
        - dinner
        - snack
    required:
    - date
    - items
    - mealType
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Fitness tracking API for personal trainers and athletes
  title: GymTrack API
  version: "1.0"
paths:
  /clients/{id}/meals:
    get:
      consumes:
      - application/json
      description: Retrieve paginated meal history for a specific client (trainers
        only, requires active relationship)
      parameters:
      - description: Client (athlete) ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Number of meals to return (default: 50)'
        in: query
        name: limit
        type: integer
      - description: 'Number of meals to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: Start date in RFC3339 format
        in: query
        name: startDate
        type: string
      - description: End date in RFC3339 format
        in: query
        name: endDate
        type: string
      - description: Filter by meal type (breakfast, lunch, dinner, snack)
        in: query
        name: mealType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Client meals retrieved successfully" "Returns meals array and
            count
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid date format
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only trainers can view client meals or no active relationship
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve meals
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get client's meals (trainer view)
      tags:
      - Meals
...
  /meals:
    get:
      consumes:
      - application/json
      description: Retrieve paginated meal history for the authenticated athlete with
        optional date range filtering
      parameters:
      - description: 'Number of meals to return (default: 50)'
        in: query
        name: limit
        type: integer
      - description: 'Number of meals to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: Start date in RFC3339 format
        in: query
        name: startDate
        type: string
      - description: End date in RFC3339 format
        in: query
        name: endDate
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Meals retrieved successfully" "Returns meals array and count
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid date format
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only athletes can list their meals
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve meals
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get athlete's meals
      tags:
      - Meals
    post:
      consumes:
      - application/json
      description: Create a new meal entry for the authenticated athlete
      parameters:
      - description: Create meal request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.CreateMealRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Meal created successfully
          schema:
            $ref: '#/definitions/gymtrack-backend_internal_domain_models.Meal'
        "400":
          description: Invalid request body or validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: User not authenticated
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Only athletes can create meals
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create meal
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a meal entry
      tags:
      - Meals
  /meals/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing meal (only within 24 hours of creation, athletes
        can only delete their own meals)
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Meal deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied or cannot delete after 24 hours
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Meal not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete meal
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a meal
      tags:
      - Meals
    get:
      consumes:
      - application/json
      description: Retrieve a meal by its ID (athletes can only access their own meals)
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Meal retrieved successfully
          schema:
            $ref: '#/definitions/gymtrack-backend_internal_domain_models.Meal'
        "403":
          description: Access denied
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Meal not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a specific meal
      tags:
      - Meals
    put:
      consumes:
      - application/json
      description: Update an existing meal (only within 24 hours of creation, athletes
        can only update their own meals)
      parameters:
      - description: Meal ID
        in: path
        name: id
        required: true
        type: string
      - description: Update meal request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.UpdateMealRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Meal updated successfully
          schema:
            $ref: '#/definitions/gymtrack-backend_internal_domain_models.Meal'
        "400":
          description: Invalid request body or validation failed
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Access denied or cannot edit after 24 hours
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Meal not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update meal
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a meal
      tags:
      - Meals
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"